requirements.txt
numpy>=1.20.0
tensorflow>=2.8.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
insightface>=0.5.0
Pillow>=8.0.0
face-recognition>=1.3.0
requests>=2.25.0
argparse>=1.4.0
tqdm>=4.60.0